Added Managers in the Logic Layer. Removed Mockup Data
I added the Manager interfaces in the common layer and the
Implementations are in the logic layer.  Controllers now have a property
that is a manager for the type of controller they are.

Mockup data has been removed as this data has been included in the
database creation script.  We should soon have working code to load
these objects from the database.