Set up a skeleton in DataLogic layer classes
Set up a skeleton for all classes in the DataLogic layer.  Equipment
manager is a mostly complete example.  Only a few additions need to be
made in the GetByBarCode, we need to add code to load the lists of
objects related to it.  Other classes should be ready to complete in a
manner similar to this class.